SaveAlbumUseCase now also searches for group members data

+++ app/src/main/java/com/vbshkn/ikollect/data/remote/api/DiscogsApi.kt
@@ -21,6 +21,6 @@ interface DiscogsApi {
 
     @GET("artists/{artist_id}")
     suspend fun getArtistDetails(
-        @Path("artist_id") artistId: Int
+        @Path("artist_id") artistId: Long
     ): Response<ArtistDetailsResponse>
 }
